reorganizing lib code
