+++ lib/matching/track_matches.py
@@ -3,9 +3,15 @@ import numpy as np
 import matplotlib.cm as cm
 import torch
 
-from lib.sg.matching import Matching
-from lib.sg.utils import (make_matching_plot,AverageTimer, read_image, frame2tensor)
-from  lib.io import generateTiles
+from thirdparty.SuperGluePretrainedNetwork.matching import Matching
+from thirdparty.SuperGluePretrainedNetwork.utils import (
+    make_matching_plot,
+    AverageTimer, 
+    read_image, 
+    frame2tensor
+)
+from  lib.utils import generateTiles
+
 torch.set_grad_enabled(False)
 
 # import matplotlib
